# swagger.yaml
swagger: "2.0"
info:
  title: Olibo League API
  description: |
    ## API de gestion de la Ligue Olibo

    Cette API permet de gérer intégralement une compétition de football : équipes, matchs,
    licences, votes, paiements, médias et bien plus.

    ### Authentification
    La plupart des endpoints nécessitent un **JWT Bearer Token**.
    Connectez-vous via `POST /api/auth/login` pour obtenir vos tokens,
    puis cliquez sur **Authorize** (🔒) et entrez : `Bearer <votre_access_token>`

    ### Rôles disponibles
    | Rôle | Description |
    |------|-------------|
    | `super_admin` | Accès total |
    | `admin_competition` | Gestion des compétitions et équipes |
    | `operator` | Gestion des matchs |
    | `referee` | Arbitre — feuilles de match et incidents |
    | `commissioner` | Commissaire de match |
    | `team_captain` | Capitaine d'équipe |
    | `coach` | Entraîneur |
    | `player` | Joueur |
    | `spectator` | Spectateur |
    | `journalist` | Journaliste |

  version: "1.0.0"
  contact:
    name: Équipe Olibo
    email: contact@olibo.com
  license:
    name: Propriétaire
host: "127.0.0.1:8000"
basePath: "/"
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "Format requis : **Bearer {access_token}**"
...
definitions:
...
  User:
    type: object
    properties:
      id:
        type: integer
        example: 1
      email:
        type: string
        format: email
        example: jean.dupont@email.com
      first_name:
        type: string
        example: Jean
      last_name:
        type: string
        example: Dupont
      phone:
        type: string
        example: "+237600000000"
      role:
        type: string
        enum: [super_admin, admin_competition, operator, referee, commissioner, team_captain, coach, player, spectator, journalist]
        example: team_captain
      is_active:
        type: boolean
        example: true
      created_at:
        type: string
        format: date-time
        example: "2026-01-15T10:30:00Z"
...
  UserCreate:
    type: object
    required: [email, password, first_name, last_name, role]
    properties:
      email:
        type: string
        format: email
        example: arbitre@olibo.com
      password:
        type: string
        minLength: 8
        example: motDePasse123
      first_name:
        type: string
        example: Paul
      last_name:
        type: string
        example: Martin
      role:
        type: string
        enum: [super_admin, admin_competition, operator, referee, commissioner, team_captain, coach, player, spectator, journalist]
        example: referee
      phone:
        type: string
        example: "+237611111111"
      is_active:
        type: boolean
        default: true
  UserUpdate:
    type: object
    properties:
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
      email:
        type: string
        format: email
      password:
        type: string
        minLength: 8
        description: Nouveau mot de passe
      old_password:
        type: string
        description: Requis si changement de mot de passe
      role:
        type: string
        description: Modifiable uniquement par un admin
      is_active:
        type: boolean
        description: Modifiable uniquement par un admin
...
  PaginatedUsers:
    type: object
    properties:
      users:
        type: array
        items:
          $ref: '#/definitions/User'
      total:
        type: integer
        example: 45
      page:
        type: integer
        example: 1
      per_page:
        type: integer
        example: 20
      pages:
        type: integer
        example: 3

  # --- Season ---
...
  Vote:
    type: object
    properties:
      id:
        type: integer
        example: 1
      member_id:
        type: integer
        example: 15
      competition_id:
        type: integer
        example: 1
      vote_type:
        type: string
        enum: [player_of_day, player_of_competition]
        example: player_of_day
      matchday:
        type: integer
        example: 5
      created_at:
        type: string
        format: date-time
  VoteResult:
    type: object
    properties:
      member_id:
        type: integer
        example: 15
      vote_count:
        type: integer
        example: 42
      vote_type:
        type: string
        example: player_of_day
      matchday:
        type: integer
        example: 5

  # --- Incident ---
...
  Error:
    type: object
    properties:
      error:
        type: string
        example: Unauthorized
      message:
        type: string
        example: Token JWT invalide ou expiré

# ============================================================
# PATHS
# ============================================================
paths:
...
  /api/users/create:
    post:
      tags: [Users]
      summary: Créer un utilisateur (réservé aux admins)
      description: Création d'un utilisateur avec n'importe quel rôle par un `super_admin` ou `admin_competition`.
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/UserCreate'
      responses:
        201:
          description: Utilisateur créé
          schema:
            $ref: '#/definitions/User'
        400:
          description: Données invalides
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Permissions insuffisantes
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Email ou téléphone déjà utilisé
          schema:
            $ref: '#/definitions/Error'
  /api/users:
    get:
      tags: [Users]
      summary: Lister tous les utilisateurs
      description: Accessible aux rôles administrateurs uniquement. Retourne une liste paginée.
      security:
        - Bearer: []
      parameters:
        - in: query
          name: page
          type: integer
          default: 1
          description: Numéro de page
        - in: query
          name: per_page
          type: integer
          default: 20
          maximum: 100
          description: Résultats par page (max 100)
        - in: query
          name: role
          type: string
          description: Filtrer par rôle
          enum: [super_admin, admin_competition, operator, referee, commissioner, team_captain, coach, player, spectator, journalist]
        - in: query
          name: is_active
          type: boolean
          description: Filtrer par statut actif/inactif
      responses:
        200:
          description: Liste paginée des utilisateurs
          schema:
            $ref: '#/definitions/PaginatedUsers'
        403:
          description: Permissions insuffisantes
  /api/users/{id}:
    get:
      tags: [Users]
      summary: Obtenir un utilisateur
      description: Un utilisateur peut voir son propre profil. Les admins peuvent voir tous les profils.
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          required: true
          type: integer
          description: ID de l'utilisateur
      responses:
        200:
          description: Profil utilisateur
          schema:
            $ref: '#/definitions/User'
        403:
          description: Accès refusé
        404:
          description: Utilisateur non trouvé
    put:
      tags: [Users]
      summary: Modifier un utilisateur
      description: |
        Un utilisateur peut modifier son propre profil.
        Seul le `super_admin` peut changer le rôle ou le statut `is_active`.
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          required: true
          type: integer
        - in: body
          name: body
          schema:
            $ref: '#/definitions/UserUpdate'
      responses:
        200:
          description: Utilisateur mis à jour
          schema:
            $ref: '#/definitions/User'
        400:
          description: Ancien mot de passe incorrect
        403:
          description: Accès refusé
        404:
          description: Utilisateur non trouvé
    delete:
      tags: [Users]
      summary: Supprimer un utilisateur
      description: Réservé au `super_admin` uniquement.
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        204:
          description: Utilisateur supprimé
        403:
          description: Permissions insuffisantes
        404:
          description: Utilisateur non trouvé
  /api/users/{id}/deactivate:
    post:
      tags: [Users]
      summary: Désactiver un utilisateur
      description: Empêche l'utilisateur de se connecter. Réservé au `super_admin`.
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        200:
          description: Utilisateur désactivé
          schema:
            $ref: '#/definitions/User'
        403:
          description: Permissions insuffisantes
  /api/users/{id}/activate:
    post:
      tags: [Users]
      summary: Activer un utilisateur
      description: Réactive un compte désactivé. Réservé au `super_admin`.
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        200:
          description: Utilisateur activé
          schema:
            $ref: '#/definitions/User'
        403:
          description: Permissions insuffisantes

  # ==================== SEASON ====================
...
  /api/voting:
    post:
      tags: [Voting]
      summary: Voter pour un joueur
      description: |
        **Règles de vote :**
        - Un vote par utilisateur par catégorie (`vote_type`)
        - Pour `player_of_day` : le champ `matchday` est obligatoire
        - Uniquement pour la compétition **active**
        - L'utilisateur ne peut pas voter pour lui-même
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [member_id, competition_id, vote_type]
            properties:
              member_id:
                type: integer
                example: 15
              competition_id:
                type: integer
                example: 1
              vote_type:
                type: string
                enum: [player_of_day, player_of_competition]
                example: player_of_day
              matchday:
                type: integer
                description: Obligatoire si vote_type = player_of_day
                example: 5
      responses:
        201:
          description: Vote enregistré
          schema:
            $ref: '#/definitions/Vote'
        400:
          description: A déjà voté ou compétition non active
          schema:
            $ref: '#/definitions/Error'
  /api/voting/results:
    get:
      tags: [Voting]
      summary: Résultats des votes
      description: Réservé au `super_admin`. Retourne les comptages de votes.
      security:
        - Bearer: []
      parameters:
        - in: query
          name: competition_id
          type: integer
        - in: query
          name: vote_type
          type: string
          enum: [player_of_day, player_of_competition]
        - in: query
          name: matchday
          type: integer
      responses:
        200:
          description: Résultats des votes
          schema:
            type: array
            items:
              $ref: '#/definitions/VoteResult'
        403:
          description: Permissions insuffisantes
  /api/voting/has-voted:
    get:
      tags: [Voting]
      summary: Vérifier si l'utilisateur a déjà voté
      security:
        - Bearer: []
      parameters:
        - in: query
          name: vote_type
          required: true
          type: string
          enum: [player_of_day, player_of_competition]
        - in: query
          name: competition_id
          required: true
          type: integer
        - in: query
          name: matchday
          type: integer
          description: Requis si vote_type = player_of_day
      responses:
        200:
          description: Statut du vote
          schema:
            type: object
            properties:
              has_voted:
                type: boolean
                example: false
  /api/voting/{id}:
    delete:
      tags: [Voting]
      summary: Supprimer un vote
      description: Réservé au `super_admin`. Décrémente automatiquement le compteur.
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        204:
          description: Vote supprimé
        403:
          description: Permissions insuffisantes

  # ==================== INCIDENT ====================
...